Full geochemical data set from Saanich Inlet 2006 to 2014

Saanich Inlet, a seasonally anoxic fjord on the coast of Vancouver Island British Columbia, is a model ecosystem for monitoring microbial community structure, function and dynamics in relation to changing levels of water column oxygen-deficiency extensible to coastal and open ocean OMZs. This data set contains nine years of monthly time series observations in Saanich Inlet encompassing water column chemistry. Chemical bottle data collected from Station S3 (-123.505, 48.59166667) in Saanich Inlet, BC, Canada. Subset of a combination of "Saanich_TimeSeries_Chemical_DATA.csv" (all variables) and "Saanich_TimeSeries_CTD_DATA.csv" (Temperature, Salinity, Density) data sets.

geochemicals

Format

A data frame with 1605 rows and 29 variables:

Longitude

Unique geographical coordinates for sampling station. Units: Decimal degrees

Latitude

Unique geographical coordinates for sampling station. Units: Decimal degrees

Cruise

Numerical identifier of individual cruises. Units: Numeric string

Date

Date of cruise. Units: YYYY-MM-DD

Depth

Sampling depth. Units: Meters

CTD_O2

Oxygen concentration calculated from CTD SBE. Units: Micromolar

PO4

Phosphate (Bran Luebbe AutoAnalyser - colorimetric). Units: Micromolar

SI

Silicate (Bran Luebbe AutoAnalyser - colorimetric). Units: Micromolar

NO3

Nitrate (Bran Luebbe AutoAnalyser - colorimetric). Units: Micromolar

Mean_NH4

Average concentration of Ammonium for each depth (fluorometric assay). Units: Micromolar

Std_NH4

Standard deviation for Ammonium average concentration. Units: NA

Mean_NO2

Average concentration of Nitrite for each depth (colorimetric assay). Units: Micromolar

Std_NO2

Standard deviation for Nitrite average concentration. Units: NA

Mean_H2S

Average concentration of Hydrogen sulfide for each depth (colorimetric assay). Units: Micromolar

Std_H2S

Standard deviation for Hydrogen sulfide average concentration. Units: NA

Cells/ml

Cell counts value quantified by flow cytometry. Units: Number of cells per millilitre (cells/mL)

Mean_N2

Average concentration of Nitrogen gas (headspace). Units: Micromolar

Std_N2

Standard deviation for Nitrogen gas. Units: NA

Mean_O2

Average concentration of Oxygen (headspace). Units: Micromolar

Std_O2

Standard deviation for Oxygen. Units: NA

Mean_CO2

Average concentration of Carbon dioxide (headspace). Units: Micromolar

Std_CO2

Standard deviation for Carbon dioxide. Units: NA

Mean_N2O

Average concentration of Nitrous oxide (headspace or automated purge-and-trap). Units: Nanomolar

Std_N2O

Standard deviation for Nitrous oxide. Units: NA

Mean_CH4

Average concentration of Methane (headspace or automated purge-and-trap). Units: Nanomolar

Std_CH4

Standard deviation for Methane. Units: NA

Temperature

CTD temperature. Units: Celsius degrees

Salinity

CTD salinity measurement at each pressure point. Units: Practical salinity unit

Density

CTD density measurement at each pressure point. Units: Sigma-theta

Source

For a brief introduction to the data, see Hallam SJ et al. 2017. Monitoring microbial responses to ocean deoxygenation in a model oxygen minimum zone. Sci Data 4: 170158. doi: 10.1038/sdata.2017.158. More detailed information on the geochemical data can be found in Torres-Beltrán M et al. 2017. A compendium of geochemical information from the Saanich Inlet water column. Sci Data 4: 170159. doi: 10.1038/sdata.2017.159 (dataset).